delete unused, consolidate
