+++ src/game/sandbox.rs
@@ -1,0 +1,28 @@
+//! This module is the box that particles are spawned in.
+
+use bevy::{prelude::*, window::PrimaryWindow};
+
+pub(super) fn plugin(app: &mut App) {
+    app.register_type::<ScreenWrap>();
+    app.add_systems(
+        Update,
+        (apply_screen_wrap)
+    );
+}
+
+#[derive(Component, Default, Reflect)]
+#[reflect(Component)]
+pub struct ScreenWrap;
+
+fn apply_screen_wrap(
+    window: Single<&Window, With<PrimaryWindow>>,
+    mut wrap_query: Query<&mut Transform, With<ScreenWrap>>,
+) {
+    let size = window.size() + 256.0;
+    let half_size = size / 2.0;
+    for mut transform in &mut wrap_query {
+        let position = transform.translation.xy();
+        let wrapped = (position + half_size).rem_euclid(size) - half_size;
+        transform.translation = wrapped.extend(transform.translation.z);
+    }
+}
